--- tests/Unit/ApproveAddToDBTest.php
@@ -1,6 +1,0 @@
-<?php
-use PHPUnit\Framework\TestCase;
-require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../src/plugins/LifePerformancesPlugin/ApproveAddToDB.php';
-
-Class ApproveAddToDBTest extends TestCase {}
