--- tests/Unit/ApproveVideoTest.php
@@ -1,6 +1,0 @@
-<?php
-use PHPUnit\Framework\TestCase;
-require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../src/plugins/LifePerformancesPlugin/ApproveVideo.php';
-
-Class ApproveVideoTest extends TestCase {}
