--- tests/Unit/CheckSubmitButtonTest.php
@@ -1,7 +1,0 @@
-<?php
-
-use PHPUnit\Framework\TestCase;
-require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../src/plugins/LifePerformancesPlugin/CheckSubmitButton.php';
-
-Class CheckSubmitButtonTest extends TestCase {}
