--- tests/Unit/GetVideoIDTest.php
@@ -1,8 +1,0 @@
-<?php
-
-use PHPUnit\Framework\TestCase;
-require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../src/plugins/LifePerformancesPlugin/GetVideoID.php';
-
-final class GetVideoIDTest extends TestCase
-{}
