+++ tests/Unit/LifePerformancesTests/CheckSubmitButtonTest.php
@@ -1,0 +1,7 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../../src/plugins/LifePerformancesPlugin/CheckSubmitButton.php';
+
+Class CheckSubmitButtonTest extends TestCase {}
