+++ tests/Unit/LifePerformancesTests/DeleteVideoFromDBTest.php
@@ -1,0 +1,7 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../../src/plugins/LifePerformancesPlugin/DeleteVideoFromDB.php';
+
+Class DeleteVideoFromDBTest extends TestCase {}
