+++ tests/Unit/LifePerformancesTests/DeleteVideoTest.php
@@ -1,0 +1,7 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../../src/plugins/LifePerformancesPlugin/DeleteVideo.php';
+
+Class DeleteVideoTest extends TestCase {}
