+++ tests/Unit/LifePerformancesTests/SubmissionFieldsTest.php
@@ -1,7 +1,7 @@
 <?php
 use PHPUnit\Framework\TestCase;
-require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../src/plugins/LifePerformancesPlugin/SubmissionFields.php';
+require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../../src/plugins/LifePerformancesPlugin/SubmissionFields.php';
 
 class SubmissionFieldsTest extends TestCase
 {
