--- tests/Unit/SendEmailToAdminTest.php
@@ -1,6 +1,0 @@
-<?php
-use PHPUnit\Framework\TestCase;
-require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../src/plugins/LifePerformancesPlugin/SendEmailToAdmin.php';
-
-Class SendEmailToAdminTest extends TestCase {}
